backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anthropic==0.21.3
anyio==3.7.1
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi>=2024.2.2
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.105.0
filelock==3.18.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.126.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.30.1
idna==3.10
openai==1.3.5
packaging==24.2
Pillow==10.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.2.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.31.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.27.0
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.24.0
aiohttp==3.9.3
